Add an MDF maker
With three languages, and more in the future, some operations can not be
done easily in the Makefile with simple Vim or sed replacements. A Forth
program is the best option to make the data format conversion by just
interpreting the source "as is"; and also it's easier to write and
maintain than complex regexp in Vim or sed scripts.